(NOBIDS) update validator deposit only every 6 hours
